// packages/aoa-maxitor/client/src/features/sidebar/types.ts
// packages/aoa-maxitor/client/src/features/sidebar/types.ts
export type NodeRow = {
  id: string;
  parent_id: string | null;
  label: string;
  type: string;
};

export type SidebarPayload = {
  level1_nodes: NodeRow[];
  level2_diagrams: NodeRow[];
  level2_nodes: NodeRow[];
  level3_diagrams: NodeRow[];
};

export type SidebarGroupedMaps = {
  l2ByParent: Map<string, NodeRow[]>;
  l3ByParent: Map<string, NodeRow[]>;
  diagramsByParent: Map<string, NodeRow[]>;
};